Making the new file a bit less polished

+++ scripts/resources/main/java/com/github/demo/model/Book.java
@@ -11,7 +11,7 @@ public class Book {
 
     private String cover;
 
-    private int rating;
+    private long rating;
 
     public Book() {
